Fix getStopDistanceObstacle and getStopDistanceStopline

+++ waypoint_planner/include/waypoint_planner/velocity_set/velocity_set_info.h
@@ -105,12 +105,12 @@ public:
     return detection_height_bottom_;
   }
 
-  int getStopDistanceObstacle() const
+  double getStopDistanceObstacle() const
   {
     return stop_distance_obstacle_;
   }
 
-  int getStopDistanceStopline() const
+  double getStopDistanceStopline() const
   {
     return stop_distance_stopline_;
   }
